split utility functions out into separate files. replaced deepObjectCompare with lodash.equals
