+++ .gitignore
@@ -7,7 +7,7 @@ bower_components
 modules
 build
 gh-pages
-lib
+/lib
 
 # vscode stuff
 .vscode
