+++ package.json
@@ -58,6 +58,7 @@
   "dependencies": {
     "element-resize-detector": "^1.1.12",
     "interact.js": "^1.2.6",
+    "lodash.isequal": "^4.5.0",
     "react-pure-render": "^1.0.2"
   },
   "peerDependencies": {
