+++ src/lib/utility/dom-helpers.js
@@ -1,0 +1,22 @@
+export function hasSomeParentTheClass (element, classname) {
+  if (element.className && element.className.split(' ').indexOf(classname) >= 0) return true
+  return element.parentNode && hasSomeParentTheClass(element.parentNode, classname)
+}
+
+export function getParentPosition (element) {
+  var xPosition = 0
+  var yPosition = 0
+  var first = true
+
+  while (element) {
+    if (!element.offsetParent && element.tagName === 'BODY' && element.scrollLeft === 0 && element.scrollTop === 0) {
+      element = document.scrollingElement || element
+    }
+    xPosition += (element.offsetLeft - (first ? 0 : element.scrollLeft) + element.clientLeft)
+    yPosition += (element.offsetTop - (first ? 0 : element.scrollTop) + element.clientTop)
+    element = element.offsetParent
+    first = false
+  }
+  return { x: xPosition, y: yPosition }
+}
+
